[Feat] #132 - 검색 Activity를 Fragment로 변경
- 기존: SearchActivity에 ~CurFrag와 ~ResultFrag가 있었음.
- 변경: 단일 Fragment로 변경
- nav_gragh에 action으로 CafeListFrag와 CafeInfoFrag 추가
- SearchFrag에서는 검색창 하단을 FrameLayout으로 만들어서 상황에 따라 RecyclerView와 ConstraintLayout 변경
- 기존에 있던 ~CurFrag와 ~ResultFrag는 삭제. xml도 삭제
