+++ app/src/main/java/org/cazait/ui/component/cafelist/CafeListFragment.kt
@@ -1,7 +1,6 @@
 package org.cazait.ui.component.cafelist
 
 import android.Manifest
-import android.content.Intent
 import android.content.pm.PackageManager
 import android.os.Build
 import android.util.Log
@@ -23,7 +22,6 @@ import org.cazait.ui.adapter.CafeListVerticalAdapter
 import org.cazait.ui.adapter.ItemDecoration
 import org.cazait.ui.base.BaseFragment
 import org.cazait.ui.component.Mapper.toCafe
-import org.cazait.ui.component.search.SearchActivity
 import org.cazait.utils.observe
 import pub.devrel.easypermissions.EasyPermissions
 import pub.devrel.easypermissions.EasyPermissions.PermissionCallbacks
@@ -111,8 +109,7 @@ class CafeListFragment : BaseFragment<FragmentCafeListBinding, CafeListViewModel
 
     private fun setSearch() {
         binding.searchBar.setOnClickListener {
-            val intent = Intent(requireContext(), SearchActivity::class.java)
-            startActivity(intent)
+            navigateToCafeSearch()
         }
     }
 
@@ -143,6 +140,10 @@ class CafeListFragment : BaseFragment<FragmentCafeListBinding, CafeListViewModel
         findNavController().navigate(
             CafeListFragmentDirections.actionCafeListFragmentToCafeInfoFragment(cafe)
         )
+    }
+
+    private fun navigateToCafeSearch() {
+        findNavController().navigate(CafeListFragmentDirections.actionCafeListFragmentToSearchFragment())
     }
 
     private fun handleHorizontalCafeList(status: Resource<FavoriteCafes>) {
