+++ app/src/main/res/layout/fragment_search.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.component.search.SearchActivity">
+        tools:context=".ui.component.search.SearchFragment">
 
         <com.google.android.material.appbar.AppBarLayout
             android:id="@+id/appBarLayout_search"
@@ -83,14 +83,79 @@
             </androidx.constraintlayout.widget.ConstraintLayout>
         </com.google.android.material.appbar.AppBarLayout>
 
-        <androidx.fragment.app.FragmentContainerView
-            android:id="@+id/search_frag"
+        <FrameLayout
             android:layout_width="0dp"
             android:layout_height="0dp"
             android:background="@color/main_black"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/appBarLayout_search" />
+            app:layout_constraintTop_toBottomOf="@+id/appBarLayout_search">
+
+            <androidx.recyclerview.widget.RecyclerView
+                android:id="@+id/rv_search"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:visibility="visible"
+                app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager" />
+
+            <androidx.constraintlayout.widget.ConstraintLayout
+                android:id="@+id/layout_search_result"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:background="@color/main_white"
+                android:visibility="gone">
+
+                <TextView
+                    android:id="@+id/text1"
+                    style="@style/TextValue_B"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:layout_marginStart="30dp"
+                    android:layout_marginTop="20dp"
+                    android:text="@string/search_text1"
+                    android:textColor="@color/main_black"
+                    android:textSize="20sp"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintTop_toTopOf="parent" />
+
+                <TextView
+                    android:id="@+id/search_text"
+                    style="@style/TextValue_B"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="@string/search_body"
+                    android:textColor="@color/main_black"
+                    android:textSize="20sp"
+                    app:layout_constraintBottom_toBottomOf="@+id/text1"
+                    app:layout_constraintStart_toEndOf="@+id/text1"
+                    app:layout_constraintTop_toTopOf="@+id/text1" />
+
+                <TextView
+                    android:id="@+id/text2"
+                    style="@style/TextValue_B"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:text="@string/search_text2"
+                    android:textColor="@color/main_black"
+                    android:textSize="20sp"
+                    app:layout_constraintBottom_toBottomOf="@+id/text1"
+                    app:layout_constraintStart_toEndOf="@+id/search_text"
+                    app:layout_constraintTop_toTopOf="@+id/text1" />
+
+                <androidx.recyclerview.widget.RecyclerView
+                    android:id="@+id/rv_search_result"
+                    android:layout_width="0dp"
+                    android:layout_height="0dp"
+                    android:layout_marginStart="10dp"
+                    android:layout_marginTop="20dp"
+                    android:layout_marginEnd="10dp"
+                    app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+                    app:layout_constraintBottom_toBottomOf="parent"
+                    app:layout_constraintEnd_toEndOf="parent"
+                    app:layout_constraintStart_toStartOf="parent"
+                    app:layout_constraintTop_toBottomOf="@+id/text1" />
+            </androidx.constraintlayout.widget.ConstraintLayout>
+        </FrameLayout>
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
